Work in progress : Extend plugin functionality * named plugins * new "getPlugin" (by name) method * Move all "auto-route" logic to dedicated plugin * Move runtime value collection/output to dedicated plugin

+++ src/Debug/Plugin/Manager.php
@@ -14,11 +14,15 @@ namespace bdk\Debug\Plugin;
 
 use bdk\Debug;
 use bdk\Debug\AssetProviderInterface;
+use bdk\Debug\ConfigurableInterface;
+use bdk\Debug\PluginInterface;
 use bdk\Debug\Plugin\CustomMethodTrait;
 use bdk\Debug\Route\RouteInterface;
 use bdk\PubSub\Event;
 use bdk\PubSub\SubscriberInterface;
 use InvalidArgumentException;
+use OutOfBoundsException;
+use RuntimeException;
 use SplObjectStorage;
 
 /**
@@ -28,13 +32,18 @@ class Manager implements SubscriberInterface
 {
     use CustomMethodTrait;
 
-    /** @var SplObjectStorage */
-    protected $registeredPlugins;
     protected $methods = array(
         'addPlugin',
+        'addPlugins',
         'hasPlugin',
+        'getPlugin',
         'removePlugin',
     );
+
+    /** @var SplObjectStorage */
+    protected $registeredPlugins;
+
+    protected $namedPlugins = array();
 
     /**
      * Constructor
@@ -45,28 +54,22 @@ class Manager implements SubscriberInterface
     }
 
     /**
-     * {@inheritDoc}
-     */
-    public function getSubscriptions()
-    {
-        return array(
-            Debug::EVENT_CUSTOM_METHOD => 'onCustomMethod',
-        );
-    }
-
-    /**
      * Extend debug with a plugin
      *
-     * @param AssetProviderInterface|SubscriberInterface $plugin object implementing SubscriberInterface and/or AssetProviderInterface
+     * @param AssetProviderInterface|SubscriberInterface $plugin Object implementing SubscriberInterface and/or AssetProviderInterface
+     * @param string                                     $name   Optional plugin name
      *
      * @return Debug
      * @throws InvalidArgumentException
      */
-    public function addPlugin($plugin)
+    public function addPlugin($plugin, $name = null)
     {
         $this->assertPlugin($plugin);
         if ($this->hasPlugin($plugin)) {
             return $this->debug;
+        }
+        if ($plugin instanceof PluginInterface) {
+            $plugin->setDebug($this->debug);
         }
         if ($plugin instanceof AssetProviderInterface) {
             $this->debug->rootInstance->getRoute('html')->addAssetProvider($plugin);
@@ -75,12 +78,73 @@ class Manager implements SubscriberInterface
             $this->addSubscriberInterface($plugin);
         }
         if ($plugin instanceof RouteInterface) {
-            $refMethod = new \ReflectionMethod($this->debug, 'onCfgRoute');
-            $refMethod->setAccessible(true);
-            $refMethod->invoke($this->debug, $plugin, false);
+            $this->addRouteInterface($plugin);
+        }
+        if (\is_string($name)) {
+            $this->namedPlugins[$name] = $plugin;
         }
         $this->registeredPlugins->attach($plugin);
         return $this->debug;
+    }
+
+    /**
+     * Add plugins defined in configuration
+     *
+     * @param array $plugins List of plugins and/or plugin-definitions
+     *
+     * @return void
+     */
+    public function addPlugins(array $plugins)
+    {
+        \array_walk($plugins, function ($plugin, $key) {
+            $cfg = array();
+            if (\is_array($plugin)) {
+                if (empty($plugin['class'])) {
+                    throw new RuntimeException(\sprintf('plugins[%s]: missing "class" value', $key));
+                }
+                $cfg = $plugin;
+                $class = $cfg['class'];
+                unset($cfg['class']);
+                $plugin = new $class();
+            } elseif (\is_string($plugin)) {
+                $plugin = new $plugin();
+            }
+            try {
+                $this->addPlugin($plugin, $key);
+            } catch (InvalidArgumentException $e) {
+                throw new RuntimeException(\sprintf('plugins[%s]: %s', $key, $e->getMessage()));
+            }
+            if ($plugin instanceof ConfigurableInterface && !empty($cfg)) {
+                $plugin->setCfg($cfg);
+            }
+        });
+    }
+
+    /**
+     * Get plugin by name
+     *
+     * @param string $pluginName Plugin name
+     *
+     * @return AssetProviderInterface|SubscriberInterface
+     *
+     * @throws InvalidArgumentException
+     * @throws OutOfBoundsException
+     */
+    public function getPlugin($pluginName)
+    {
+        if (\is_string($pluginName) === false) {
+            throw new InvalidArgumentException(\sprintf(
+                'getPlugin expects a string. %s provided',
+                $this->debug->php->getDebugType($pluginName)
+            ));
+        }
+        if ($this->hasPlugin($pluginName) === false) {
+            throw new OutOfBoundsException(\sprintf(
+                'getPlugin(%s) - no such plugin',
+                $this->debug->php->getDebugType($pluginName)
+            ));
+        }
+        return $this->namedPlugins[$pluginName];
     }
 
     /**
@@ -92,6 +156,9 @@ class Manager implements SubscriberInterface
      */
     public function hasPlugin($plugin)
     {
+        if (\is_string($plugin)) {
+            return isset($this->namedPlugins[$plugin]);
+        }
         $this->assertPlugin($plugin);
         return $this->registeredPlugins->contains($plugin);
     }
@@ -99,12 +166,22 @@ class Manager implements SubscriberInterface
     /**
      * Remove plugin
      *
-     * @param SubscriberInterface $plugin object implementing SubscriberInterface
+     * @param string|SubscriberInterface $plugin Plugin name or object implementing SubscriberInterface
      *
      * @return Debug
      */
-    public function removePlugin(SubscriberInterface $plugin)
+    public function removePlugin($plugin)
     {
+        if (\is_string($plugin)) {
+            $plugin = $this->removePluginNamed($plugin);
+        }
+        if ($plugin === false) {
+            return $this->debug;
+        }
+        $pluginName = \array_search($plugin, $this->namedPlugins, true);
+        if ($pluginName !== false) {
+            unset($this->namedPlugins[$pluginName]);
+        }
         $this->registeredPlugins->detach($plugin);
         if ($plugin instanceof AssetProviderInterface) {
             $this->debug->rootInstance->getRoute('html')->removeAssetProvider($plugin);
@@ -116,6 +193,23 @@ class Manager implements SubscriberInterface
     }
 
     /**
+     * Remove plugin by name
+     *
+     * @param string $pluginName Plugin name
+     *
+     * @return AssetProviderInterface|SubscriberInterface|false The removed plugin instance, or false
+     */
+    private function removePluginNamed($pluginName)
+    {
+        if (isset($this->namedPlugins[$pluginName]) === false) {
+            return false;
+        }
+        $plugin = $this->namedPlugins[$pluginName];
+        unset($this->namedPlugins[$pluginName]);
+        return $plugin;
+    }
+
+    /**
      * Add SubscriberInterface plugin
      *
      * @param SubscriberInterface $plugin SubscriberInterface instance
@@ -124,11 +218,10 @@ class Manager implements SubscriberInterface
      */
     private function addSubscriberInterface(SubscriberInterface $plugin)
     {
-        $this->debug->eventManager->addSubscriberInterface($plugin);
         $subscriptions = $plugin->getSubscriptions();
         if (isset($subscriptions[Debug::EVENT_PLUGIN_INIT])) {
             /*
-                plugin we just added subscribes to Debug::EVENT_PLUGIN_INIT
+                plugin subscribes to Debug::EVENT_PLUGIN_INIT
                 call subscriber directly
             */
             \call_user_func(
@@ -137,6 +230,29 @@ class Manager implements SubscriberInterface
                 Debug::EVENT_PLUGIN_INIT,
                 $this->debug->eventManager
             );
+        }
+        $this->debug->eventManager->addSubscriberInterface($plugin);
+    }
+
+    /**
+     * Add RouteInterface plugin
+     *
+     * @param RouteInterface $route RouteInterface instance
+     *
+     * @return void
+     */
+    private function addRouteInterface(RouteInterface $route)
+    {
+        $classname = \get_class($route);
+        $prefix = 'bdk\\Debug\\Route\\';
+        $containerName = 'route' . \substr($classname, \strlen($prefix));
+        if (\strpos($classname, $prefix) === 0 && isset($this->debug->{$containerName}) === false) {
+            $this->debug->setCfg('serviceProvider', array(
+                $containerName => $route,
+            ));
+        }
+        if ($route->appendsHeaders()) {
+            $this->debug->obStart();
         }
     }
